Update scipy_optimize_chi2.py
Finish both methods in scipy to minimize chi2. The first is linear-regression to find a starting c0 and then find a minimized c nearby (locally, because it is assumed that these two minimal points are near) 2. use a global minimal method.

Also, the constraint of cs sum to 1 is done by constraint in optimization instead of dividing by thier sum. The later one will cause a inconsistancy in the value of c before and after the finction is called.